width and height as size_t

+++ grid.h
@@ -27,7 +27,7 @@ typedef struct
  * @param dims the grid view dimensions.
  * @return int
  */
-inline int width(GridDims *dims)
+inline size_t width(GridDims *dims)
 {
     return dims->col_end - dims->col_start + 2;
 }
@@ -38,7 +38,7 @@ inline int width(GridDims *dims)
  * @param dims the grid view dimensions.
  * @return int
  */
-inline int height(GridDims *dims)
+inline size_t height(GridDims *dims)
 {
     return dims->row_end - dims->row_start + 2;
 }
